+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add java class find-definition support, decompiling .class files when available. #762 
+- Add `:java :download-jdk-source?` setting to download JDK source after startup if not cached before globally. Disabled by default, enable if you want to hava java support for JDK classes.
 
 ## 2022.03.26-18.47.08